Harden order checkout after code review
Review findings applied: map the versioned cart-delete race to 409
ORDER_CONFLICT instead of a 500 through the global catch-all; widen the
order total to BIGINT/long so a large cart cannot silently overflow;
pin line-item ordering with @OrderBy; move the empty-cart invariant
into Order.place as the single source (it now throws EmptyCartException
directly); bind the snapshot product once in OrderItem.from.

+++ src/main/resources/db/migration/V5__create_orders.sql
@@ -7,7 +7,7 @@ CREATE TABLE orders
 (
     id          BIGINT      NOT NULL AUTO_INCREMENT,
     user_id     BIGINT      NOT NULL,
-    total_price INT         NOT NULL,
+    total_price BIGINT      NOT NULL,
     created_at  DATETIME(6) NOT NULL,
     PRIMARY KEY (id),
     KEY idx_orders_user (user_id),
